refactor(hello-work-job-searcher): layout/skin分離とBEM風クラス命名によるCSS構造整理
