+++ apps/hello-work-job-searcher/src/app/Layout.module.css
@@ -1,17 +1,24 @@
-.layoutTitle {
-  text-align: left;
+.header {
+  flex: 1;
+}
+
+.header-title {
+  text-align: left;          /* 配置 */
+  padding: 0.7em 0 0.9em 0.5em; /* 部品の骨格 */
+  border-bottom: 1px solid #ccc; /* 骨格の境界線 */
+  box-shadow: none;          /* 骨格の影の有無 */
+}
+
+.header-title--primary {
   font-weight: 700;
   font-size: 1.35rem;
-  padding: 0.7em 0 0.9em 0.5em;
   color: #f3f4f6;
   font-family: 'Merriweather', 'Times New Roman', Times, serif;
   letter-spacing: 0.04em;
   background: #4b5563;
-  border-bottom: 1px solid #ccc;
-  box-shadow: none;
 }
 
-.layoutRoot {
+.wrapper {
   display: flex;
   flex-direction: column;
   height: 100vh;
@@ -19,11 +26,7 @@
   box-sizing: border-box;
 }
 
-.layoutHeader {
-  flex: 1;
-}
-
-.layoutMain {
+.main {
   flex: 9;
   overflow: auto;
 }
