+++ apps/hello-work-job-searcher/src/app/_components/Job/index.module.css
@@ -1,0 +1,31 @@
+.headerSection {
+  min-height: 0;
+  flex: 1;
+}
+
+.detailSection {
+  min-height: 0;
+  flex: 9;
+}
+
+
+.layout-job-detail {
+  padding: 1em;
+}
+
+
+.layout-job-detail--primary {
+  border: 0.1em solid gainsboro;
+}
+
+
+.layout-job-detailList {
+  display: flex;
+  flex-direction: column;
+  gap: 1em;
+}
+
+
+.layout-job-detailList--primary {
+  list-style: none;
+}
