+++ apps/hello-work-job-searcher/src/app/_components/index.css
@@ -14,18 +14,6 @@
   height: 100%;
 }
 
-.job-detail {
-  border: 0.1em solid gainsboro;
-  padding: 1em;
-}
-
-.job-detail ul {
-  list-style: none;
-  display: flex;
-  flex-direction: column;
-  gap: 1em;
-}
-
 .modal-backdrop {
   position: absolute;
   top: 0;
