+++ apps/hello-work-job-searcher/src/app/jobs/@detail/[jobNumber]/layout.tsx
@@ -1,11 +1,12 @@
+import styles from "./page.module.css";
 export default function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
   return (
-    <div>
-      <div>{children}</div>
+    <div className={styles["layout-container"]}>
+      {children}
     </div>
   );
 }
