+++ apps/hello-work-job-searcher/src/app/jobs/@detail/[jobNumber]/page.tsx
@@ -1,7 +1,6 @@
 import { jobFetchSuccessResponseSchema } from "@sho/models";
 import * as v from "valibot";
 import { JobDetail } from "@/app/_components/Job";
-import styles from "./page.module.css";
 
 interface PageProps {
   params: Promise<{ jobNumber: string }>;
@@ -29,12 +28,6 @@ export default async function Page({ params }: PageProps) {
   };
 
   return (
-    <main className={styles.mainSection}>
-      <div className={styles.layoutContainer}>
-        <div className={styles.detailSection}>
-          <JobDetail jobDetail={jobDetail} />
-        </div>
-      </div>
-    </main>
+    <JobDetail jobDetail={jobDetail} />
   );
 }
