+++ apps/hello-work-job-searcher/src/app/jobs/@search/page.module.css
@@ -1,33 +1,71 @@
-.mainSection {
+/* Layout / Structure */
+
+.layout-search {
   height: 100%;
+  display: flex;
+  flex-direction: column;
 }
 
-.splitLayoutContainer {
+.layout-search-header {
+  flex: 1;
+  /* min-height: 0; */
+}
+
+.layout-search-list {
+  flex: 9;
+  overflow: scroll;
+}
+
+
+
+/* ページ全体の左右分割 */
+.layout-root {
   display: flex;
   height: 100%;
+  gap: 1.2rem;
+  /* flex間の余白 */
 }
 
-.jobDetailLayoutContainer {
-  flex: 1;
-}
-
-.searchLayoutContainer {
+/* 左右カラム */
+.layout-column {
   flex: 1;
   display: flex;
   flex-direction: column;
   height: 100%;
-  gap: 1.2rem;
 }
 
-
-
-.searchLayoutHeaderSection {
+/* Search カラム内の縦分割 */
+.layout-column-header {
   min-height: 0;
+  /* flexbox 内で overflow:auto の子要素が正しく働くため */
 }
 
-
-
-.searchLayoutListSection {
-  min-height: 0;
+.layout-column-list {
   flex: 1;
+  min-height: 0;
+  /* 残りスペースを占有 */
+}
+
+/* Optional: JobDetail カラムも同じ構造として使える */
+.layout-job-detail {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+}
+
+/* Parts（骨格だけ） */
+.header-title {
+  padding: 0.7em 0 0.9em 0.5em;
+  border-bottom: 1px solid #ccc;
+  box-shadow: none;
+}
+
+/* Skin（見た目） */
+.header-title--primary {
+  font-weight: 700;
+  font-size: 1.35rem;
+  text-align: left;
+  font-family: 'Merriweather', 'Times New Roman', Times, serif;
+  letter-spacing: 0.04em;
 }
