+++ apps/hello-work-job-searcher/src/app/styles/globals.css
@@ -1,6 +1,8 @@
-html, body {
-    height: 100vh;
-    overflow: hidden;
+html,
+body {
+    margin: 0;
+    padding: 0;
+    height: 100%;
 }
 
 h1 {
